requirements.txt
PyPDF2
python-docx
langchain
langchain-ollama
groq
gradio
